Move Result to the jsonrpc package next to where its used
