+++ client/src/main/java/com/lmax/solana4j/client/jsonrpc/Result.java
@@ -1,8 +1,8 @@
-package com.lmax.solana4j.client.util;
+package com.lmax.solana4j.client.jsonrpc;
 
 import java.util.function.Function;
 
-public final class Result<L, R>
+final class Result<L, R>
 {
     private final L error;
     private final R success;
@@ -13,37 +13,37 @@ public final class Result<L, R>
         this.success = success;
     }
 
-    public static <L, R> Result<L, R> error(final L error)
+    static <L, R> Result<L, R> error(final L error)
     {
         return new Result<>(error, null);
     }
 
-    public static <L, R> Result<L, R> success(final R success)
+    static <L, R> Result<L, R> success(final R success)
     {
         return new Result<>(null, success);
     }
 
-    public boolean isError()
+    boolean isError()
     {
         return error != null;
     }
 
-    public boolean isSuccess()
+    boolean isSuccess()
     {
         return success != null;
     }
 
-    public L getError()
+    L getError()
     {
         return error;
     }
 
-    public R getSuccess()
+    R getSuccess()
     {
         return success;
     }
 
-    public <S> Result<L, S> onSuccess(final Function<R, Result<L, S>> onSuccess)
+    <S> Result<L, S> onSuccess(final Function<R, Result<L, S>> onSuccess)
     {
         if (isSuccess())
         {
